Pyflyte meta inputs (#1823)
* Re-orgining pyflyte run


* Pyflyte beautified and simplified


* fixed unit test


* Added Launch options


* lint fix


* test fix


* fixing docs failure


---------
