+++ tests/flytekit/unit/cli/pyflyte/test_register.py
@@ -7,7 +7,7 @@ from click.testing import CliRunner
 
 from flytekit.clients.friendly import SynchronousFlyteClient
 from flytekit.clis.sdk_in_container import pyflyte
-from flytekit.clis.sdk_in_container.helpers import get_and_save_remote_with_click_context
+from flytekit.clis.sdk_in_container.helpers import get_and_save_remote_with_click_context, get_remote
 from flytekit.configuration import Config
 from flytekit.core import context_manager
 from flytekit.remote.remote import FlyteRemote
@@ -36,6 +36,13 @@ t = ShellTask(
         script="echo 'Hello World'",
     )
 """
+
+
+@mock.patch("flytekit.clis.sdk_in_container.helpers.FlyteRemote")
+def test_get_remote(mock_remote):
+    r = get_remote(None, "p", "d")
+    assert r is not None
+    mock_remote.assert_called_once_with(Config.for_sandbox(), default_project="p", default_domain="d")
 
 
 @mock.patch("flytekit.clis.sdk_in_container.helpers.FlyteRemote")
